analytics/requirements.txt
flask==1.1.2
flask_login==0.5.0
flask_wtf==0.14.3
python-decouple==3.4
gunicorn==20.0.4
